+++ module-graph/src/test/java/ch/hsr/adv/ui/graph/logic/GraphStringifyerTest.java
@@ -2,7 +2,7 @@ package ch.hsr.adv.ui.graph.logic;
 
 import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.ui.core.access.FileDatastoreAccess;
-import ch.hsr.adv.ui.core.logic.util.ADVParseException;
+import ch.hsr.adv.ui.core.logic.exceptions.ADVParseException;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.inject.Inject;
